This commit adds: -The ability to see ticket notes -css changes- -getting updated list of entities

+++ client/src/utils/options.js
@@ -1,7 +1,7 @@
 export let options = {
-  weekday: "long",
+  weekday: "short",
   year: "numeric",
-  month: "long",
+  month: "short",
   day: "numeric",
   minute: "2-digit",
   hour: "numeric"
